agentic_ai/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv
pypdf
duckduckgo-search
google-generativeai
